Refactor to separate remote and local serializer factories
